refactor: move nodejs functions to own file
