+++ src/pfe/portal/modules/metricsService/node.js
@@ -1,0 +1,65 @@
+const fs = require('fs-extra');
+const path = require('path');
+const util = require('util');
+
+const Logger = require('../utils/Logger');
+const { deepClone } = require('../utils/sharedFunctions');
+
+const log = new Logger(__filename);
+
+const getPathToPackageJson = (projectDir) => path.join(projectDir, 'package.json');
+const getPathToBackupPackageJson = (projectDir) => path.join(projectDir, 'backupPackage.json');
+
+async function injectMetricsCollectorIntoNodeProject(projectDir) {
+  const pathToPackageJson = getPathToPackageJson(projectDir);
+  const originalContentsOfPackageJson = await fs.readJSON(pathToPackageJson);
+
+  const pathToBackupPackageJson = getPathToBackupPackageJson(projectDir);
+  await fs.writeJSON(pathToBackupPackageJson, originalContentsOfPackageJson, { spaces: 2 });
+
+  const newContentsOfPackageJson = getNewContentsOfPackageJson(originalContentsOfPackageJson);
+  log.trace(`Injecting metrics collector into project's package.json, which is now ${ util.inspect(newContentsOfPackageJson) }`);
+
+  await fs.writeJSON(pathToPackageJson, newContentsOfPackageJson, { spaces: 2 });
+}
+
+async function removeMetricsCollectorFromNodeProject(projectDir) {
+  const pathToBackupPackageJson = getPathToBackupPackageJson(projectDir);
+  const backupPackageJson = await fs.readJSON(pathToBackupPackageJson);
+
+  const pathToPackageJson = getPathToPackageJson(projectDir);
+  await fs.writeJSON(pathToPackageJson, backupPackageJson, { spaces: 2 });
+  log.trace(`Restored project's package.json to ${ util.inspect(backupPackageJson) }`);
+
+  await fs.remove(pathToBackupPackageJson);
+}
+
+function getNewContentsOfPackageJson(originalContentsOfPackageJson) {
+  const newContentsOfPackageJson = deepClone(originalContentsOfPackageJson);
+
+  newContentsOfPackageJson.scripts.start = getNewStartScript(originalContentsOfPackageJson.scripts.start);
+  newContentsOfPackageJson.dependencies['appmetrics-codewind'] = '^0.1.0';
+  return newContentsOfPackageJson;
+}
+
+function getNewStartScript(originalStartScript) {
+  const metricsCollectorScript = '-r appmetrics-codewind/attach';
+
+  if (originalStartScript.includes(metricsCollectorScript)) {
+    return originalStartScript;
+  }
+
+  const splitOriginalStartScript = originalStartScript.split(' ');
+  const indexOfNodeCmd = splitOriginalStartScript.findIndex(word => ['node', 'nodemon'].includes(word));
+
+  let newStartScript = deepClone(splitOriginalStartScript);
+  newStartScript.splice(indexOfNodeCmd + 1, 0, metricsCollectorScript);
+  newStartScript = newStartScript.join(' ');
+
+  return newStartScript;
+}
+
+module.exports = {
+  injectMetricsCollectorIntoNodeProject,
+  removeMetricsCollectorFromNodeProject,
+};
